[BENCH] change weights from uint to floats

+++ 00_graph_bench/include/algorithms/cachesim/SSSP.h
@@ -42,8 +42,8 @@ void freeSSSPStats(struct SSSPStats *stats);
 // ********************************************************************************************
 uint32_t SSSPAtomicMin(uint32_t *dist, uint32_t new);
 uint32_t SSSPCompareDistanceArrays(struct SSSPStats *stats1, struct SSSPStats *stats2);
-int SSSPAtomicRelax(uint32_t src, uint32_t dest, uint32_t weight, struct SSSPStats *stats);
-int SSSPRelax(uint32_t src, uint32_t dest, uint32_t weight, struct SSSPStats *stats);
+int SSSPAtomicRelax(uint32_t src, uint32_t dest, float weight, struct SSSPStats *stats);
+int SSSPRelax(uint32_t src, uint32_t dest, float weight, struct SSSPStats *stats);
 void durstenfeldShuffle(uint32_t *vertices, uint32_t size);
 
 // ********************************************************************************************
